check for nil peer and update unit tests
